fix(app): Debugging Setting/Getting/Saving Data
Example Application now simulates a behavior, as we would have
in CoreMedia Studio. General speaking, we want to prevent semantically
equivalent data triggering an auto-checkout of a document in
CoreMedia Studio.

To do so, we must ensure, that the data forwarded to server
are strictly equivalent (`===`). In Studio we will solve this by
remembering CKEditor's document version directly on `setData` and
on autosave possibly forward these set data rather than the transformed
ones by CKEditor, if the version did not change yet.