fix(content-link): Must Accept `undefined` as ContentUriPath
In 3dc5383160c0f0647e12152a12587b50bf4b9915 we hardened the type-guard
`hasContentUriPath` – according to the interface of `HasContentUriPath`.

But: On remove/cancel of a content-link the value of `hasContentUriPath`
becomes `undefined` rather than `null`. Thus, we have to handle `undefined`
in the same way as `null`.

Adjusting this type-check we now _know that we are responsible for taking
care of the (unset) value_, which again opens doors for continue processing
content-link removal.

fixes: #83