docs: Straighten Packages List and Provide Installation Hints
The packages list was outdated, wrong and incomplete. Adjusted
the list and merged them to one list, as this is easier to maintain
and better to scan for any possibly missing packages.

Added a small section on _Installation_ telling, what the
essential plugins are about.

Added a section about the Studio Integration, which especially
handled the possible challenge regarding unexpected auto-checkout
of contents and how to prevent this.