test(playwright): Calm Jest regarding open handles
We should wait for the server to finish eventually, by waiting
for the callback to return. Otherwise, we see this in the console:

> A worker process has failed to exit gracefully and has been force
> exited. This is likely caused by tests leaking due to improper
> teardown. Try running with `--detectOpenHandles` to find leaks.
> Active timers can also cause this, ensure that `.unref()` was
> called on them.

Note, that the warnings is still logged, but `--detectOpenHandles`
does not complain anymore on the unclosed server. Last try, the
report was empty. Before, fixing, it complained about the unclosed
server but also regarding the service-agent dependency.