build(eslint): Migrate to JavaScript Configuration
JavaScript provides some more options when it comes to specific
parser options to pass. Thus, migrating to JavaScript configuration
for ESLint.

Also overhauled the configuration to extension specific configurations
and applying all recommended rules with overrides very required.
Some overrides got added as "broken window", which is, that we
may want to address them later, but not now.