feat(styles): Styles Feature: Font Sizes
Activating Styles feature for font-size classes, which were
available prior to CMCC 11.2210.1. Note, that it may not be
a good approach using the Styles feature here for various
reasons, like not being able to "increase" a font size by
some toolbar button. Thus, it may be better to first search
for alternative plugins for given classes.