refactor(richtext): Introduce Module ckeditor5-dom-support
The new data conversion will be based on plain DOM. For this,
we are going to provide a bunch of utilities to interact with
the DOM like, for a first example, to straighten type guards
for DOM objects.

As first example, a type-guard is added for `ParentNode`, which,
as it is an interface, cannot simply be detected via
`instanceof` like for the objects the type-guard is based on.