refactor(dom-converter): Tests towards production use-cases
Testing HtmlDomConverter with special focus on production use-cases,
we require later on, which is, for example, dealing with adjusting
table structures.

Along with that, added a bunch of convenience API especially for
tables, as they provide rather poor support for tables originating
from CoreMedia Rich Text 1.0, where table elements are not
represented as HTMLTableElement, thus, do not provide convenience
API for accessing several nested table elements such as
`<thead>` or `<tfoot>`.