fix(link): Prevent open link in same tab if in read only.
In read only mode the link balloon does not open but instead the URL will be opened in same tab.
Kind of annoying for users if they accidentally click the link and lose context.
For Content Links a new tab in the work area is opened. For external links the URL is opened in a new browser tab.