refactor(data-processor): First Mappings for XLink Attributes
For images and anchors we require some more sophisticated mapping
for XLink Attributes, like merging xlink:role and xlink:show into
target attribute for anchors.

Nevertheless, with this new code, we have some fallback to persist
the remaining properties.

Note, that as `dataset` does some name checks, we had to migrate
from `-` in key name to `_`. This also requires to adapt some
existing code (not completed yet), as well as the GHS-registration.