refactor(data-processor): Adding Sanitizer
Adding first version of sanitizer, that is not aware of
attributes yet.

Added another set of rules for _suppressed elements_:
These are elements, which are send to data view by
CKEditor 5, but which we want to ignore. This is
<figure> element currently. This may need to be
adapted, if the figure element holds any relevant
data for us.