refactor(data-processor): Fixes xml:lang Attribute during Sanitation
`setAttribute("xml:lang", "en")` creates an attribute
node with `namespaceURI` set to `null` and the `localName`
set to `xml:lang` (and prefix to `null`). Thus, to be able
to properly handled the attribute during sanitation and
possibly other stages, we need to ensure, that we always
use the `*NS` methods.