refactor(data-processor): Better `appended` for DocumentFragments
When a `DocumentFragment` is appended, it dissolves to its
child nodes at the parent node, while the fragment itself
is emptied.

Having this, invoking `appended(parent, fragment)` would
provide no robust option to access the actual child nodes
appended to parent as part of this process.

Instead, applied special handling in HtmlDomConverter, that
triggers extra reporting for child nodes of fragments.