refactor(data-processor): Straighten Reporting
Up to now, no signal `appended` got raised when attaching
to the target document eventually. As corresponding methods
to raise a signal are protected API, decided to add a new
method `convertAndAppend` to DomConverter API. Now rules
could also listen to the final append to the target document
and trigger corresponding actions.