refactor(itest): Should Fix Playwright Tests
Playwright tests needed to listen to when the data
were set, to continue with the test. The corresponding
typing needed to be adjusted (although naively, also
the old typing should still work, as both data-processors,
old and new are Emitters).

We may also think about migrating typing to `Emitter`, but
skipped that for now, especially, as binding is more explicit
this way.

Note, that currently tests cannot be executed locally on
Windows, though.