fix(dom-converter): Fixes xmlns="" on element import
Elements got imported with `xmlns=""` at least observable
in Firefox. Reason was again, that `lookupNamespaceURI(null)`
does not provide the expected result. Having this, we fell
into the same pit, as already documented in `Elements.createElement`:
To prevent empty `xmlns` attribute, we must ensure using the same
`namespaceURI` as the document.