fix(dom-converter): Split Custom Types and TypeGuards
For internal re-use within ckeditor5-dom-support, we need
to extract TypeGuards and Types, as otherwise, we get cyclic
references.

Still, for external use, considered public API to be part of
dedicated modules `Elements`, `ParentNodes`, ... While this
for some utilities provide rather empty modules despite some
export, it is more consistent and easier to extend this way
(so the assumption).