fix(images): Catch unwanted xlink-href attributes
We cannot assume that processed images have a valid uriPath as xlink-href.
In case the attribute is not a valid uriPath, we must exit early, to not run
into an error which would break the editor.

We Therefore must check if the attribute is set on downcast, right before creating the spinner marker.
If the xlink-href is empty (empty string), simply show a broken image.
(Skip the rest of the computation)

Later, the xlink-href attribute is converted to the resulting uriPath.
If this fails, do not break the editor, but instead just return early.
In this case, the spinner will stay.