fix(link): Register Other Link Attributes
You may now register other link attributes, which are
not yet handled by corresponding plugins, to belong
to a link, thus, may not be used outside of a link.

Having this, we can cleanup these attributes when
removing the link and we can handle cursor
positioning (e.g., two step move caret) correctly.