refactor(link): Create ckeditor5-link-common
Similar to ckeditor5-core-common we now have a
module dedicated to CKEditor 5 Link Feature.
Corresponding attribute handling is now located
in this new module as well as access to the
Link UI.

This eases to prevent dependency cycles, as we
are going to use these utilities in
ckeditor5-coremedia-richtext, too.