refactor(link): Move xlink:type support to Data Processor
`xlink:type` mapped to `data-xlink-type` is a fixed attribute
in CoreMedia Rich Text 1.0 DTD. Thus, we don't expect any plugin
providing editing features for `data-xlink-type`. Having this,
it is safe (and considered best practice), that the data processor
plugin takes over and registers it as belonging to a link.

Not doing so for `title` and `data-xlink-actuate`, as someone
may provide a plugin in the future, taking care of these. While
they should still use `LinkAttributes` from their plugins to
register handling these, it should be hard-wired in existing
plugins. Instead, extended the README.md to provide details
how to handle this (and an example is available in
ckeditor.ts).