fix(data-processor): Support `title` attribute for anchors
As with previous data-processor (v10), we now map the
`xlink:title` attribute to `title`.

While in previous data-processor we did so also for images,
we won't go that way for now, as it collides with the title
attribute being used only for editing view purpose. Thus,
we had a collision for img.title anyway.

Instead, GRS needs to be configured to keep `data-xlink-title`
for images.