fix(data-processor): Fix title-support for images
In previous and current data-processor we had issues with
the xlink:title-attribute. It collided with the design
decision using the `title` attribute as provider for a
tooltip in editing view.

This commit straightens the behavior, so that at least
`xlink:title` information is not lost. Still, we need
to decide at some point in the future, how to resolve
the conflict - thus, as soon as we want to offer
editing features for `xlink:title` for image elements.