feat(link): Single Letter Content Name
We observed, that trying to open a content whose
name consists only of a single letter does not
work. A click to open triggers no action.

This commit adds an example of such a document to
the drag and drop examples. As single letter
character, we have chosen the unicode symbol
U+1F54A (🕊️) to possibly add another challenge
to the link dialog.