refactor(font-mapper): Introduce Augmentation Approach for EditorConfig, PluginsMap
Adapting to the recent refactorings at CKEditor, we should/must update our
typings, so that configuration options can be correctly resolved. This requires
to introduce `index.ts` and, following the pattern provided by CKEditor, a
file `augmentation.ts`.

Note the effect, for example in `FontMapper`, where this line does not require
a cast or explicit type anymore:

```typescript
const customFontMapperConfig = config.get(COREMEDIA_FONT_MAPPER_CONFIG_KEY);
```

It is automatically detected to be of type `FontMapperConfig | undefined`.

Similar to CKEditor, suggestion is, that `index.ts` is carefully designed to
denote our public API. We may also think of replacing `index-doc.ts`. Then we
need to decide, if it is ok, that there are less classes mentioned in
`index.ts` than in `index-doc.ts` (which currently contains all classes).

Not yet resolved: This approach does not work for usages such as in
`ckeditor.ts`. Failed to access `index.ts` successfully in there.