refactor(build): Required upgrade of xml-formatter
We are about to adjust `tsconfig.json` to support `index.ts`
import of our own artifacts. This not only enables but forces
us to update xml-formatter from 2.6.1 to 3.3.2, as we need
proper support in TypeScript.

Before, we had to revert the upgrade, as it was incompatible
with our module setup, as it seems.