refactor(jest): Prefer ts-jest over Babel
As it seems, Babel provides many irrelevant issues trying
to transform TypeScript on the fly as part of the Jest
Configuration.

`ts-jest` does a much better job here. Still, Babel seems
to be required for the CKEditor 5 dependencies, to deal
with the JavaScript artifacts.

With the current state, all Jest (Unit) tests run again
without any issues.