fix(docs): TypeDoc failed with the new setup
Rewrote configuration to align with the proposed example at
https://github.com/Gerrit0/typedoc-packages-example.

One benefit: No more extra management of all paths to be
included.

I stick to `index-doc.ts` for now, as our "file of truth"
for the files to be included into the documentation. We
may want to review later, if to use `index.ts` instead.