fix: Fixes Link-Activation for Caret-Movement by Cursor-Keys Central element of this change is to fix a false-negative `isEnabled` state of the `OpenInTabCommand`. It failed to update the enabled state on caret-movement via cursor-keys.
Also, some logging has been added to follow and understand
decisions, for example, on enabled states.

**Refactoring details:**

# ckeditor5-coremedia-content

## OpenInTabCommand

Refactored to "base" command for any OpenInTab Command. As can be seen
in CKEditor 5 link-editing and image-feature, there are some different
approaches to bind commands to the actual model element.

To benefit from "copy & adapt" from CKEditor 5 sources, we decided to
make the implementation more specific to the actual use-case by
extending the command rather by configuring it.

In addition to these changes, the `execute()` method may now be
invoked with explicit URI paths to open instead of relying on
the model state. Note though, that this only works, if the command
is enabled, as a high priority listener prohibits execution, if
the command is disabled.

Also, the `execute` methods provides a return value now, as this
is what commands could do according to the API. This opens a
chance to callers, to wait for the command to finish (unused yet).

Another important fix is to set `affectsData` to `false`. As this
command's enabled state is not bound to the read-only state of
the editor, it should be set from default `true` to `false`.

## OpenInTab

Extracts common utility methods to interact with the WorkAreaService,
like to check of contents can be opened or to trigger opening them.

As we see for other changes, there is no need to check in advance,
if a content can be opened. Contents, that could not be opened will
just be returned as "rejected". This is one reason, why the
`OpenContentInTabCommand` is now always enabled (see below for
details).

# ckeditor5-coremedia-images

This module got adapted according to the refactorings of
`OpenInTabCommand` now having a dedicated `OpenImageInTabCommand`.

## ContentImageOpenInTabUI

Adapted to new command. Also removed the `isEnabled` check, as
this is already done implicitly by `execute` and the corresponding
high-priority listener installed in `Command` to prevent invocation
of `execute` if the command is disabled.

## OpenImageInTabCommand

This is the new command only dedicated to opening image/BLOB
documents. It benefits from the suggested "copy & adapt" approach
to identify related images. Here, it is similar to, for example,
`ResizeImageCommand` provided by CKEditor 5.

It also handled the enabled state according to the availability/
readability of the referenced content. This is, because the enabled
state is also bound to the button in UI.

Still, we enable the button by default, even if we don't know if
the content is readable yet, as a subsequent call to open the content
would reject opening the content anyway. Thus, we prefer short latency
for the 90% case while editing contents.

## ContentImageEditingPlugin

Using the new command now.

# ckeditor5-coremedia-link

Similar to the image plugin, we now have an extra command to open
linked contents within this module. Different to the image plugin,
as we don't use the enabled state of the command yet to represent
it in the UI, we don't update this state but default to always
enabled as all commands by default.

## ContentLinkActionsViewExtension

Using the new command now. As we know the URIs to open already, there
is no need to get the URI from the model state. Thus, we invoke the
command with the URI path we already know.

## ContentLinkViewFactory

Adapted to new command, just as `ContentLinkActionsViewExtension`.
Now also using the already known URI path, instead of guessing it
from model.

## ContentLinks

Using the new command and moved instantiation of command to the
`init` phase. There is no need to instantiate the command again
and again (probably, without deep code analysis).

## OpenContentInTabCommand

Implementation of `OpenInTabCommand` dedicated for use in
content link feature.

Note, that the command refreshes its value, but is ignored for now
from given usages mentioned above (they provide the URI path
directly).

We could have also decided to not use a command but just to
open the contents directly. Benefit of the command: We have
an extra layer, where we may add to disable/enable the command
and thus, prevent opening contents at a central place.

Also, the idea is, that we could use the enabled state to bind it to
the corresponding buttons in the UI. This again would require to
carefully track the model state and to add listeners, if a linked
content becomes unreadable meanwhile (destroyed, or moved to an
unreadable location).