build(rimraf): Prepare Build to src/
We decided to switch to building to src/ as this seems to be
common practice and does not trigger error-prone auto-complete
when importing in IDEs, that sometimes suggest the dist/ folder
as the folder to add to the import statement.

Also, we hope, that we get closer to the recommended artifact
setup, as it ships with CKEditor 5 core plugins.

As first step, adapted `rimraf` to respect the new structure,
while still providing an portal to the old setup (just to
clean up our own workspaces, or while still migrating to new
structure).

Also, removed `rimraf` from dependencies but now aligned all
invocations to `npx`, as we already did for removing node_modules.