feature(bbcode-dataprocessor): Take child contents into account in HTML2BBCodeRule
Instead of returning bbcode with the node's textContent, we should return the child contents instead.
This is mandatory since a strong/italic etc. node might have children that need to be rendered inside the
resulting bbcode tag. Since we are computing the bbcode string recursively, starting with the inner nodes,
we can just use the already computed child string.