feature(bbcode-dataprocessor): Fix falsy EditorConfig types
The EditorConfig type extensions in some of our augmentation files were marked as mandatory.
This is false. TypeScript would complain about insufficient configs if these plugins were not configured.
(Which is totally fine) Therefore, these config extensions are now marked as optional.

Also, the configuration of the CoreMedia RichText plugin was now moved into a separate function.
TypeScript was not happy with compatibility and rules being set independently because these values
have to match in order to represent a certain config type (v10 or latest). Compatibility is now
also set explicitly to make sure the value matches the given type. (type expects "v10" or "latest" and
the value was of type "string | true")