feat: Refactor To Delegating Plugin
Now, we have one plugin, that delegated to the corresponding
API. For now, we always use the "highest level" API, which is
`ContextAwareCachedDataAccess`. If required, we could add some
configuration option here.

Note, that while we do not use the intermediate layers of the
data-access API, it helps getting an overview of the different
responsibilities of the different layers.