feat: Introduce preProcessAnchorElement and mapArtificialXLinkRole
The main feature is `mapArtificialXLinkRole` which is required, if you
used `xlink:role` to store other link-related states, especially
when `xlink:show` is not equal to `other`. While for `xlink:show`
set to `other` `xlink:role` is meant to represent the `target`, usage
is not specified for any other value of `xlink:show`.

Using the new `mapArtificialXLinkRole` enables to override the default
behavior of also writing these artificial `xlink:role` information to
`target` attribute in view layers.