refactor: Refactor Read Only Toggle
Now also using the SwitchButton. We decided to extend the SwitchButton,
so that it ships with the option to delay switching on demand, to cover
a relevant use-case for switching read-only mode (we require the delay,
so that we may, for example, open some dialogs and see, that they
respond for a switch of the read-only mode).