feat: RgbColor and CSSStyleDeclaration Utilities
Introducing a new API for accessing colors in DOM, meant especially for
`toData` processing (from data view to data).

While its primary focus (and usage) for now is dedicated to the BBCode
`[color]` tag, it may be of value also for mapping color styles to
CoreMedia Rich Text 1.0, encoded as tokens in class-attribute, for
example.