refactor: Dive into BBob API
To address https://github.com/JiLiZART/BBob/issues/125 diving somewhat
deeper into the BBob API. Assumption is, that the best place to
respect paragraphs is the `process` method, as we require dedicated
handling of text nodes and must not take them as is.

The current state is more of "inlining the HTML5-Preset", so that we
may start with experiments.