feat: Add Support For Paragraphs on Root
To be able to process paragraphs on root-level we need to create
an artificial root-node to be processed within configured
default-tags.

There may be other options, like adding a paragraph-support plugin,
that runs after the HTML5 preset. But the current solution also
fixes an issue with the default `quote` processing of BBob HTML5
Preset: By default, it just wraps any content into a single paragraph
not respecting, that in a valid HTML5 structure, a paragraph must not
contain other block-level elements.

As alternative, we can now use the new "intelligent paragraph
processing" also on this level already, instead of having to devour
the results of the HTML5 preset (which even may change with more
recent BBob versions).