fix: Skip Empty Paragraphs
One of the original tests for BBob HTML5 Preset (we copied them, to
ensure, we are consistent with the original mapping), added spaces
in front of a list. These extra spaces triggered creation of a
paragraph node containing only these _indent spaces_.

Decided to introduce a `skipEmpty` option, that is enabled by default.
This will prevent the creation of paragraphs, that are empty - also
trimming any blanks prior to this check.