test: Integration Test for Unique Attribute Behavior
Adding an integration test for behavior influenced by:

* https://github.com/JiLiZART/BBob/issues/202

We decided (better: will decide with subsequent commits) to accept
the possible flaw as corner case, that a URL tag like

```
[url=onclick]T[/url]
```

will still get the `onclick` value removed, as it is denoted in
as `onclick=onclick` in attributes (concept of unique attributes as
it is now in BBob).

If these tests fail, we have to revisit corresponding code and see,
if we need or can adapt something.