refactor: Unique Attribute Handling in HtmlSanitzer
Straightened the behavior for unique attribute handling and made it
an explicit option of the configurable predicate. This way, anyone
could veto our decision to handle unique attributes just as all others
regarding filtering of possibly malicious attributes.

Also added a bunch of documentation (also internal documentation for
developers) to clearly describe the behavior of the sanitizer.