test: Broken Tests Due To Sanitizer
The sanitizer seems to run "too early", i.e., it does not know of any
magic attributes generated from content. This again only applies to
URLs such as completely valid BBCode:

```
[url]https://example.org/?A=1&B=2[/url]
```

Running the sanitizer as early plugin, we will escape the content
naively just to:

```
[url]https://example.org/?A=1&amp;B=2[/url]
```

The issue: In HTML5 Preset (or our derived solution) the content
defaults to the `href` attribute, which again is automatically
escaped within BBob processing.

This will cause as final rendering result in current setup:

<a href="https://example.org/?A=1&amp;amp;B=2">https://example.org/?A=1&amp;B=2</a>