test: More Possibly Broken Rendering Results
Adding more tests around the [code] tag handling, that demonstrates
misbehavior to some context:

    [code][i]T[/i][/code]

will be transformed to:

    <pre><code class="language-plaintext"><i>T</i></code></pre>

which means, that the `[i]` tag is not processed via the configured
rules but an as-is copy. This is bad.

Similar to this, but here we benefit from disallowing unknown tags, we
see another effect for:

    [code][script]javascript:alert("X")[/script][/code]

which is processed to:

    <pre><code class="language-plaintext">[script]javascript%3Aalert(&quot;X&quot;)[/script]</code></pre>

This feels like rather unpredictable behavior, denoted to the current
workaround to get out of too deep recursions by "early rendering".