test: Adapt CKEditor 5 Preset Test to Flawed Unique Attr. Behavior
As we now "keep all other attributes" for `[url]` and do not only
rely on the unique attribute, we dive deeper into the issues of
the referenced BBob issue.

Again, as soon as the corresponding issue got fixed, we assume, that
we may need to adapt the code - or at least remove these now irrelevant
tests.