fix: Fix Babel Configuration
The previous configuration had issues with object-destructuring,
for example. The following code produced two objects spread1
and spread2 having the very same attributes, while they should
differ regarding the override order:

```
(() => {
  const key: string = "key";
  const uniqueAttrValue = "override";
  const otherAttrs = { [key]: "original" };
  const spread1 = { [key]: uniqueAttrValue, ...otherAttrs};
  const spread2 = { ...otherAttrs, [key]: uniqueAttrValue};

  console.log("probe", {
    spread1,
    spread2,
  });
})();
```

We now just use the defaults as is and rely on the presets to come
with the required plugins.