fix(blocklist): Reduce risk of running into trouble with marker name limitations
Storing the blocked word in the marker name has the pitfall that marker names may have limitations which are hard to foresee.
One limitation is that `,` are not allowed in marker names. A sentence containing `,` would cause a console error and can't be blocked.
Therefore, store the blocked word with the id as a reference.