feat: Pave `[img]` support
Up to now, `[img]` got stripped, because we did not mark it as
"allowed" BBCode tag. This is, because we had no toData-processing,
yet.

It exists now, with this commit. To benefit from some CKEditor 5
features and to provide some first "named attribute handling" in
our BBCode, added support for `alt` attribute, as supported by
some BBCode dialects.