feat: Rework and Finish Font Size Mapping
There is a wide variety of possible interpretations of `[size=42]`.
CKEditor 4 BBCode assumed it to be a percentage. Many BBCode dialects
assume some ordinal value representing an enumeration entry such as
tiny, small, big.

None of the references I found takes size as "size in pixels" (which
does not necessarily mean, that this is not a valid interpretation).

Because of this, and because the CKEditor 5 Font Size Plugin only
supports either font-size given in pixels (as numeric values) or
font-size implicitly defined by classes, reverted the logic to
represent an enumeration.

This solution covers many aspects:

* It is very comparable to the CKEditor 4 BBCode plugin based
  on percentages.

* It matches many references I found, where some even just migrated
  to an enumeration based approach.

* It prevents possibly malicious font-sizes to reach the editing view,
  which is font-sizes that would just "blow up" the editor or
  font-sizes that are barely visible in the UI.

The mapping is as follows (relying on the CKEditor 5 Font Size defaults
regarding `toView` class and `em` mapping:

|  Input Range  | `toView` Class | Suggested `em` mapping | `toData` Normalization |
|:-------------:|----------------|-----------------------:|-----------------------:|
|  0 ≤ N < 78   | `text-tiny`    |                `0.7em` |                     70 |
|  78 ≤ N < 93  | `text-small`   |               `0.85em` |                     85 |
| 93 ≤ N < 120  | _none_         |                  `1em` |                    100 |
| 120 ≤ N < 160 | `text-big`     |                `1.4em` |                    140 |
|    160 ≤ N    | `text-huge`    |                `1.8em` |                    180 |