refactor: Remove [table] Support
The [table] support is too flawed and requires much more work, to
robustly integrate into the expectations of CKEditor 5 (e.g., when and
how table headings are used).

We decided to skip that feature for now, to possibly revoke it later on.
If we do, we should think about options (similar as we do in
Rich Text Data Processor), how to ensure the expected structure by
CKEditor 5 (column headers need to be placed into a thead section, as
one example).

We may also think about supporting common table cell options like
rowspan and colspan, so that it integrates more nicely into
CKEditor 5.

On the other hand: Who would actually write tables in BBCode? Solutions
found for Markdown, for example, having a representation also readable
in plain text seem to be the much better options nowadays.