docs: Add Reference to data-facade
The main README.md describes issues with the Autosave feature and
"unexpected changes" without editorial interaction.

Meanwhile, we have the data-facade plugin, that provides support for
this aspect. Decided to keep the original description, as it is still
true, if you want to solve the issue on your own, but added a reference
to the data-facade plugin, that may be used instead.