ci: Fix Semantic Versioning For Pre-Release
There is a strict specification for pre-release versions.

Previous template `<mj>.<mn>.<p>-pr<pr>-<run>.0` violated this pattern
and may irritate parsing and ordering semantic versions.

The new template aligns with the standards, so that, for example, a
previous pre-release version:

    16.0.1-pr158-717.0

is now generated as:

    16.0.1-pr.158.717.0

See-also: <https://semver.org/>