build: Add Script to Validate peerDependencies
Adding script that may be used to validate our peerDependencies. This
is meant to handle often observed issues, that the `peerDependencies`
are not aligned with the actual dependencies.

As of now, the check will only fail, if we switch to fixed versions
for CKEditor 5, rather than using version ranges. As this is
recommended anyway (you must not mix CKEditor 5 package versions
within a deliverable), it is possibly an option to switch to fixed
version references, just as done in CKEditor 5 packages.