build: Remove Engine Requirements
We do not need to specify engine requirements in our deliverables and
if we do, we would require to use a Node.js version that is compatible
with all consumers of our API.

Thus, removing engine requirements from all our `packages/`. Skipped
removing it in `app/` and `itest/` as this may require some further
guessing, if it may be removed. As both are not delivered, there is no
issue keeping the engine-requirements in `package.json`.