refactor: First Steps To Support ImageBlock
This is a preliminary commit showing some sketches to also support
the ImageBlock plugin. Much more is to be done (like data-processing),
but we are also blocked by our way of stripping the "src" attribute from
the model, which blocks switching the image type from ImageInline to
ImageBlock and vice versa.

Either, we keep some artificial non-empty src attribute in model, or
move the removal of the artificial attribute to the dataDowncast layer,
or, best option, CKEditor 5 opens the ImageTypeCommand to accept more
attributes than just `src` and `uploadId` as denoting a
"valid non-broken image".