refactor: Remove `preventUpcastImageSrc`
There is no need, that we remove the `src` attribute from `<img>` during
upcast. It is enough, that we have (at least) a late processing in
data-processing, that ensures corresponding `src` attribute to be
stripped.

Nevertheless, it may make sense removing the `src` attribute already
earlier in `dataDowncast`, so that there is no need for the
data-processor to take core of it. Leaving the new state as is would
cause some subtle breaking change for anyone who wrote a custom
rich text data-processor.