fix(link): Harden Default Target Calculation
The post-fixer for links is also invoked when removing links. In this
case, the calculated link is `null`.

The previous code struggled, because brute-force casting to `string`
instead of doing some type-checking.

Also, the previous code was not robust, when any of the filters got an
issue. As such filters may be part of the configuration, we should be
robust on invocation and handle it in some predictable way.

The fix is two-fold:

* We do not invoke the filter on link removal anymore.
* We added a robustness layer for custom filters, if they fail with
  exception. Decision is to ignore the corresponding filter, which is,
  that on failure, no default target will be selected.