fix(content-clipboard): Harden Multi-Content Drop
When dropping multiple content items, it may happen, that we are at
the start or end of the document. If doing so, we saw an exception like:

    DataToModelMechanism: Error occurred in promise CKEditorError: writer-split-element-no-parent

While this was an issue before, it seems to become more severe with
update to CKEditor 5 40.2.0. Now we check the required precondition,
that a required parent element is available - and skip splitting,
otherwise.

It may be questioned, if the start/end checks are still required then,
or if we have a double check here. Did not answer the question, but
felt comfortable, that this change fixes the issue.