feat(coremedia-richtext)!: Align `setXLinkDataSetEntries` with `setXLinkAttributes`
As `setXLinkAttributes` and `setXLinkDataSetEntries` are kind of
meant to "work as a pair", both should provide a comparable
experience.

* Added `allowEmpty` as just done for `setXLinkAttributes`.
* Added TSdoc for the method

Note, that this is a breaking change, as, different from which the
line-comment suggested, this method did not ignore empty
values. Now it does (by default).

All usages should be reviewed and possibly adapted.

BREAKING CHANGE: `setXLinkDataSetEntries` now by default ignores entries with empty values.