refactor(coremedia-richtext)! Redecide on `allowEmpty` default
In many use-cases it makes more sense not to ignore empty
values for XLink entries. Such as for attributes that are required
(`xlink:href` for `<a>` and `<img>`).

This reverts the previous decision to "break" the API of
`setXLinkDataSetEntries`. Instead, the breaking change now
applies to `setXLinkAttributes`.

Still, the goal is, to keep both aligned.

`allowEmpty = true` as default is also considered "more harmless"
as the result is less surprising, if you invoke this API. If empty
values are ignored, you may be surprised, if you did not read the
TSdoc (or code) before.

BREAKING CHANGE: `setXLinkAttributes` may need to set `allowEmpty = true` explicitly. Reverting previous breaking change for `setXLinkDataSetEntries`.